tests/operator: split lifecycle and config tests into separate files
Move test Describe blocks from the monolithic operator.go into two
thematic files:

- lifecycle.go: reconciliation, component configuration, imagePullSecrets,
  infrastructure management, monitoring (PrometheusRule/ServiceMonitor),
  adopt-unmanaged-entities, and RoleAggregationStrategy tests. These all
  need the full KV restabilization teardown.

- config.go: node placement, replicas, certificate rotation, feature gates,
  seccomp, instancetypes, virt-template, and external CA tests.

Both files use a BeforeEach to initialise per-test state and a matching
AfterEach to restore the KubeVirt installation after each (potentially
destructive) test, mirroring the structure of the original operator.go.
deprecatedBeforeAll is removed since all callers are gone.

operator.go now holds only shared helpers, constants, and the SIGSerial
wrapper — no Describe blocks remain.
